requirements.txt
alembic==1.4.3
bcrypt==3.2.0
boto3==1.17.16
botocore==1.20.16
cffi==1.14.3
click==7.1.2
commonmark==0.9.1
Faker==4.15.0
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
flask-marshmallow==0.14.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.9.1
marshmallow-sqlalchemy==0.24.0
mdpdf==0.0.9
psycopg2==2.8.6
pycparser==2.20
PyMuPDF==1.18.9
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
s3transfer==0.3.4
six==1.15.0
SQLAlchemy==1.3.20
text-unidecode==1.3
urllib3==1.26.3
Werkzeug==1.0.1
